removed cart OneToOne and placed only in account to avoid recursive code

+++ src/main/java/it/develhope/shoppyz/order/OrderService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface OrderService {
 
-        public Optional<Order> getOrder(Integer id);
+        public Optional<Order> getOrder(Long id);
         public void deleteOrder(Order order);
         public void postOrder(Order order, Account account);
         public void updateOrder(Order order);
